chore(lottery-v2): Remove mention of v2 apart from in contracts
