+++ src/App.tsx
@@ -18,7 +18,7 @@ import history from './routerHistory'
 // Only pool is included in the main bundle because of it's the most visited page
 const Home = lazy(() => import('./views/Home'))
 const Farms = lazy(() => import('./views/Farms'))
-const LotteryV2 = lazy(() => import('./views/LotteryV2'))
+const Lottery = lazy(() => import('./views/Lottery'))
 const Ifos = lazy(() => import('./views/Ifos'))
 const NotFound = lazy(() => import('./views/NotFound'))
 const Collectibles = lazy(() => import('./views/Collectibles'))
@@ -60,7 +60,7 @@ const App: React.FC = () => {
               <Pools />
             </Route>
             <Route path="/lottery">
-              <LotteryV2 />
+              <Lottery />
             </Route>
             <Route path="/ifo">
               <Ifos />
