+++ src/views/Lottery/components/AllHistoryCard.tsx
