+++ src/views/Lottery/components/BuyTicketsButton.tsx
