+++ src/views/Lottery/components/BuyTicketsModal/BuyTicketsModal.tsx
