+++ src/views/Lottery/components/BuyTicketsModal/TicketInput.tsx
