+++ src/views/Lottery/components/BuyTicketsModal/generateTicketNumbers.tsx
