+++ src/views/Lottery/components/BuyTicketsModal/useTicketsReducer.tsx
