+++ src/views/Lottery/components/CheckPrizesSection.tsx
