+++ src/views/Lottery/components/ClaimPrizesModal/index.tsx
