+++ src/views/Lottery/components/HowToPlay.tsx
