+++ src/views/Lottery/components/PreviousRoundCard/Footer.tsx
