+++ src/views/Lottery/components/RewardBracketDetail.tsx
