+++ src/views/Lottery/components/RewardBrackets.tsx
