+++ src/views/Lottery/components/TicketNumber.tsx
