+++ src/views/Lottery/components/ViewTicketsModal/index.tsx
