+++ src/views/Lottery/helpers.tsx
