+++ src/views/Lottery/hooks/useGetNextLotteryEvent.ts
