+++ src/views/Lottery/pageSectionStyles.tsx
