+++ src/views/Lottery/svgs/Balls.tsx
