+++ src/views/Lottery/svgs/PoolAllocationChart.tsx
