+++ src/views/Lottery/svgs/TicketPurchaseCard.tsx
