+++ src/views/Lottery/svgs/WhiteBunny.tsx
