+++ src/views/Lottery/svgs/index.tsx
